+++ Sources/Parley/Data/PollingService.swift
@@ -55,21 +55,19 @@ final class PollingService: PollingServiceProtocol {
         didSet { timer?.invalidate() }
     }
 
-    func startRefreshing() async {
-        await MainActor.run {
-            setTimer(interval: .twoSeconds)
-            addObservers()
-        }
+    @MainActor
+    func startRefreshing() {
+        setTimer(interval: .twoSeconds)
+        addObservers()
     }
 
-    func stopRefreshing() async {
-        await MainActor.run {
-            timer?.invalidate()
-            loopRepeated = 0
-            timerInterval = .twoSeconds
-            timer = nil
-            removeObservers()
-        }
+    @MainActor
+    func stopRefreshing() {
+        timer?.invalidate()
+        loopRepeated = 0
+        timerInterval = .twoSeconds
+        timer = nil
+        removeObservers()
     }
 
     private func setTimer(interval: TimerInterval) {
@@ -83,10 +81,8 @@ final class PollingService: PollingServiceProtocol {
     private func refreshFeed() async {
         guard
             let id = await messagesManager.lastSentMessage?.remoteId,
-            timer?.isValid == true else
-        {
-            return
-        }
+            timer?.isValid == true
+        else { return }
         
         do {
             let messageCollection = try await messageRepository.findAfter(id)